Menus
So I made a start menu that has sound effects for the buttons which
light up when mouse over. The start button (activate) should take you
into the "GameScene" where all the action will occur. The how to play
button (debriefe) will take you into an instructions page that tells you
how to play and what you need to do. Both of these menus have an awesome
background song. That I think we can use for actual game play since it's
decent in length. The instructions page has a button that should bring
you to the "GameScene". The game over menu has two buttons. one that
should quit the game, and one that should take you back into the
"GameScene" as well as a burning skeleton laughing and big a GAME OVER.
HOWEVER! The skeleton always displays over the text for some reason. I
will try and resolve this tomorrow. I also made a text file that we can
list borrowed resources in.