Victory Scene
Jon reminded me that there needs to be a victory scene once the player
reaches the goal. I have most of it gone right now, since for the most
part it's a copy of the gameover scene. When I get home in a few hours I
will finish it and push the other changes